Ensure overlaps_with_... and intersect_with are consistent
It makes sense to view an overlap occuring if the intersection is not
empty.

The same should also happen to the special case of never() and
eternity() ranges. It should return False given that intersect_with
returns an empty (never()) range.

This also resulted in fixes to the tests for
is_contiguous_with_... where a range is never().

sem-ver: feature